requirements.txt
openai
python-dotenv
pypdf
chromadb
numpy
fastapi
uvicorn
